修袭击过度bug，黄金锅bug，广播对话bug，任务现可隐藏

+++ src/main/java/com/p1nero/smc/entity/custom/npc/start_npc/StartNPC.java
@@ -446,7 +446,7 @@ public class StartNPC extends SMCNpc {
     public void handleNpcInteraction(ServerPlayer player, byte interactionID) {
         DialogueComponentBuilder dialogueComponentBuilder = new DialogueComponentBuilder(this);
         SMCPlayer smcPlayer = SMCCapabilityProvider.getSMCPlayer(player);
-        double moneyRate = smcPlayer.getLevelMoneyRate();
+        int moneyRate = (int) smcPlayer.getLevelMoneyRate();
         //购买
         if (interactionID == 1) {
             if (smcPlayer.getMoneyCount() < 100 * moneyRate) {
